MISC: Add git hooks
We added three different hooks in order to verify if the commit looks correct.
These are the new hooks:
- pre-commit
- commit-msg
- pre-push

These hooks have a new function:

1. pre-commit:
Makes sure the code itself looks correct. For example Linting, licensing, unit testing is succesfull.

2. commit-msg:
Validate the commit message itself.
This looks for a section that is known in the header, for example MISC: , TDE: , LIB-TDE, SETTINGS: etc
Additionally it also looks for a description in the commit (which is mandatory)
Lastly it checks to see if the header is short enough.

3. pre-push:
Check if the user is working in the correct branch (we don't want users to push to master directly)