PLUGIN: Added desktop widget plugin
Added a new example plugin that listens for resources and draws pretty
widgets directly on the desktop.

These are to show off the extensibility of the plugin system
It includes a nice to use config file where you can easily declare how
the plugin should behave and whart type of charts to draw.