TDE: modify rules and layout
Some rules were not working properly
New options have been added to the rules to ignore/allow titlebars
Each application can tell us to ignore the titlebar or not.

Additionally a new fairh layout has been added.
This layout allows for horizontal tiling, which is usefull when using a
virtual keyboard (used when in tablet mode or when using accessability
applications)