TDE: proper support for onscreen keyboards
When you have a laptop/tablet with `TDE` you often want a virtual
keyboard. This patch allows the Onboard keyboard to work properly with
`TDE` It does this by never giving focus to onboard, instead the
previous application keeps the focus.

Note: a fix should be found to make this patch work with all virtual
keyboards, currently it only works with onboard