REFACTOR: correctly calculate thread count
Previsouly we used the siblings attribute of the first cpu core, however
this doesn't always show the correct amount of threads.
By counting each individual processor attribute we can calucalte the
thread count.
However, as for the cpu core count we only find the value of cpu cores
for single cpu systems.

When you have a multi cpu system (for example a supercomputer) then we
still only return the core count of the first detected cpu.