TDE: route all audio through the volume api
Instead of manually calling functions and commands all over the place,
abstract it into the volume api of lib-tde.

This way we have more flexibility, also we provided a new `disable_software_volume_control` option in the general settings.
This option can disable software control of the audio stack (thus
delegating it to hardware such as an amp or dac)