MISC: Fix a lot of spelling mistakes
A lot of spelling mistakes where found (using a spellchecker)
This patch adds a simple script checking spelling of comments/markdown
files.

Note that the spellcheck only checks comments in lua, not the effective
code (That would require more work)

Currently the spellchecker is terribly inefficient.
Maybe someone wants to improve it :-)