BUG: Fixes #113 - don't trigger the break timer on fullscreen
When viewing a movie/video or working focused (often
fullscreen/maximized) we don't want to show the break timer as that can
happen at the worst moment.

Imaging gaming or other apps that require focus and suddenly a break
timer popup appears. That would be very annoying.
Instead this patch only shows the break timer when needed