LIB-TDE: Add initial api for screen resolution and refresh rate
This initial api gives you the ability to see/get a list of refresh
rates/resolutions.

Before the full merger into production code this part of the api needs
to be documented.

Additionally a UI/Widget interface should be added to let the user
select the refresh rate and resolution