BUG: Don't adjust the position of clients in the auto_hide wibox
When the wiboxes were activated the client locations updated to 'fit'
the available screen, this causes clients to redraw their content.

However when repainting the client content this means buttons and other
ui elements move around the screen, causing bad UX for the user.

This patch draws the boxes over the client instead of ontop, to ensure
that the ui elements don't move around the screen.