BUG: Resolution change will fit the screens correctly
Previously when changing the resolution and using a multi-head setup,
you would run into issue's where the screens didn't get repositioned
relative to eachoter.

This could cause the screen's to overlap, or leave a enourmous DMZ
between the screens caused by the resolution change.
Now we reposition the screens relative to the other screens based on
their new resolution